gfxaccel: Metal-backed DSp/RAVE/GL/NQD acceleration + compositor, rendering fixes, logging
Squashed integration of the gfxaccel work branch.

- gfxaccel subsystem: DrawSprocket, RAVE, OpenGL/AGL/GLU, and NQD/QuickDraw
  engines on a shared Metal device + compositor (SheepShaver/src/gfxaccel),
  with Swift lifecycle services (DSp events/idle timer, background and
  memory-warning observers).
- Rendering fixes: Cro-Mag GL camera/depth + GLU bridge, UT RAVE
  gamma/mipmaps/resource-table, DSp front-buffer/AltBuffer + CLUT/gamma,
  matrix-stack depth.
- Core/runtime: 64-bit Mach exception handling, aspect-ratio preservation,
  prefs/JIT defaults, ppc stfiwx, assorted Mac Catalyst fixes.
- gfxaccel logging: unified os_log control plane with verbose tier,
  ship-default off, tools/gfxlog.sh toggle + GFXACCEL_LOG env, docs.

+++ SheepShaver/src/include/cpu_emulation.h
@@ -83,7 +83,7 @@ static inline uint64 ReadMacInt64(uint32 addr) {return *(uint64 *)addr;}
 static inline void WriteMacInt16(uint32 addr, uint32 w) {*(uint16 *)addr = w;}
 static inline void WriteMacInt32(uint32 addr, uint32 l) {*(uint32 *)addr = l;}
 static inline void WriteMacInt64(uint32 addr, uint64 ll) {*(uint64 *)addr = ll;}
-static inline uint32 Host2MacAddr(uint8 *addr) {return (uint32)addr;}
+static inline uint32 Host2MacAddr(uint8 *addr) {return (uint32)(uintptr_t)addr;}
 static inline uint8 *Mac2HostAddr(uint32 addr) {return (uint8 *)addr;}
 static inline void *Mac_memset(uint32 addr, int c, size_t n) {return memset(Mac2HostAddr(addr), c, n);}
 static inline void *Mac2Host_memcpy(void *dest, uint32 src, size_t n) {return memcpy(dest, Mac2HostAddr(src), n);}
